171034183 - Sendgrid email error for Burnside
This change sets the `page_size` to the MAX per the sendgrid API docs (1000 entries). Functionally this should support our distribution list needs for some time as its unlikely a single list would have 1001 or more entries. This should also catch an exception thrown directly by the sendgrid sdk upon invocation; i strongly suggest all applications of that sdk are updated for more defensive coding like this.

caveat: i've not tested this patch out yet...but it looks right...

This is considered a hotfix to be cherry picked against `master` if warranted.